+++ confignode/src/main/java/org/apache/iotdb/confignode/persistence/node/NodeInfo.java
@@ -33,6 +33,7 @@ import org.apache.iotdb.confignode.consensus.request.write.datanode.RemoveDataNo
 import org.apache.iotdb.confignode.consensus.request.write.datanode.UpdateDataNodePlan;
 import org.apache.iotdb.confignode.consensus.request.write.statistics.UpdateLoadStatisticsPlan;
 import org.apache.iotdb.confignode.consensus.response.DataNodeConfigurationResp;
+import org.apache.iotdb.confignode.manager.node.heartbeat.NodeStatistics;
 import org.apache.iotdb.rpc.TSStatusCode;
 import org.apache.iotdb.tsfile.utils.ReadWriteIOUtils;
 
@@ -85,9 +86,6 @@ public class NodeInfo implements SnapshotProcessor {
   private final AtomicInteger nextNodeId = new AtomicInteger(-1);
   private final Map<Integer, TDataNodeConfiguration> registeredDataNodes;
 
-  // Node Statistics
-  private final Map<Integer, NodeStatistics> nodeStatisticsMap;
-
   private final String snapshotFileName = "node_info.bin";
 
   public NodeInfo() {
@@ -96,8 +94,6 @@ public class NodeInfo implements SnapshotProcessor {
 
     this.dataNodeInfoReadWriteLock = new ReentrantReadWriteLock();
     this.registeredDataNodes = new ConcurrentHashMap<>();
-
-    this.nodeStatisticsMap = new ConcurrentHashMap<>();
   }
 
   /**
@@ -395,11 +391,6 @@ public class NodeInfo implements SnapshotProcessor {
     }
   }
 
-  /** Only used when the ConfigNode-Leader is switched */
-  public Map<Integer, NodeStatistics> getNodeStatisticsMap() {
-    return nodeStatisticsMap;
-  }
-
   @Override
   public boolean processTakeSnapshot(File snapshotDir) throws IOException, TException {
     File snapshotFile = new File(snapshotDir, snapshotFileName);
@@ -423,8 +414,6 @@ public class NodeInfo implements SnapshotProcessor {
       serializeRegisteredConfigNode(fileOutputStream, protocol);
 
       serializeRegisteredDataNode(fileOutputStream, protocol);
-
-      serializeNodeStatistics(fileOutputStream);
 
       fileOutputStream.flush();
 
@@ -464,14 +453,6 @@ public class NodeInfo implements SnapshotProcessor {
     }
   }
 
-  private void serializeNodeStatistics(OutputStream outputStream) throws IOException {
-    ReadWriteIOUtils.write(nodeStatisticsMap.size(), outputStream);
-    for (Map.Entry<Integer, NodeStatistics> nodeStatisticsEntry : nodeStatisticsMap.entrySet()) {
-      ReadWriteIOUtils.write(nodeStatisticsEntry.getKey(), outputStream);
-      nodeStatisticsEntry.getValue().serialize(outputStream);
-    }
-  }
-
   @Override
   public void processLoadSnapshot(File snapshotDir) throws IOException, TException {
 
@@ -497,8 +478,6 @@ public class NodeInfo implements SnapshotProcessor {
       deserializeRegisteredConfigNode(fileInputStream, protocol);
 
       deserializeRegisteredDataNode(fileInputStream, protocol);
-
-      deserializeNodeStatistics(fileInputStream);
 
     } finally {
       configNodeInfoReadWriteLock.writeLock().unlock();
@@ -530,17 +509,6 @@ public class NodeInfo implements SnapshotProcessor {
     }
   }
 
-  private void deserializeNodeStatistics(InputStream inputStream) throws IOException {
-    int size = ReadWriteIOUtils.readInt(inputStream);
-    while (size > 0) {
-      int nodeId = ReadWriteIOUtils.readInt(inputStream);
-      NodeStatistics nodeStatistics = new NodeStatistics();
-      nodeStatistics.deserialize(inputStream);
-      nodeStatisticsMap.put(nodeId, nodeStatistics);
-      size--;
-    }
-  }
-
   public static int getMinimumDataNode() {
     return minimumDataNode;
   }
@@ -549,7 +517,6 @@ public class NodeInfo implements SnapshotProcessor {
     nextNodeId.set(-1);
     registeredDataNodes.clear();
     registeredConfigNodes.clear();
-    nodeStatisticsMap.clear();
   }
 
   @Override
@@ -559,12 +526,11 @@ public class NodeInfo implements SnapshotProcessor {
     NodeInfo nodeInfo = (NodeInfo) o;
     return registeredConfigNodes.equals(nodeInfo.registeredConfigNodes)
         && nextNodeId.get() == nodeInfo.nextNodeId.get()
-        && registeredDataNodes.equals(nodeInfo.registeredDataNodes)
-        && nodeStatisticsMap.equals(nodeInfo.nodeStatisticsMap);
+        && registeredDataNodes.equals(nodeInfo.registeredDataNodes);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(registeredConfigNodes, nextNodeId, registeredDataNodes, nodeStatisticsMap);
+    return Objects.hash(registeredConfigNodes, nextNodeId, registeredDataNodes);
   }
 }
